feat: added build flavors (dev, stage, prod) and get_it dependency injection

+++ dvm_app/lib/features/vitals/data/models/sensor_data_model.dart
@@ -6,14 +6,12 @@ import '../../domain/entities/sensor_data.dart';
 part 'sensor_data_model.freezed.dart';
 part 'sensor_data_model.g.dart';
 
-/// Data model for SensorData
-/// Extends the domain entity and adds JSON serialization
 @freezed
 sealed class SensorDataModel with _$SensorDataModel {
   const SensorDataModel._();
 
   const factory SensorDataModel({
-    @JsonKey(name: 'thermal_value') required double thermalValue,
+    @JsonKey(name: 'thermal_value') required int thermalValue,
     @JsonKey(name: 'battery_level') required double batteryLevel,
     @JsonKey(name: 'memory_usage') required double memoryUsage,
     required DateTime timestamp,
@@ -22,13 +20,12 @@ sealed class SensorDataModel with _$SensorDataModel {
   factory SensorDataModel.fromJson(Map<String, dynamic> json) =>
       _$SensorDataModelFromJson(json);
 
-  /// Create from native platform response
   factory SensorDataModel.fromPlatform(Map<dynamic, dynamic> map) {
     return SensorDataModel(
-      thermalValue: (map['thermal_value'] as num).toDouble(),
+      thermalValue: (map['thermal_value'] as num).toInt(),
       batteryLevel: (map['battery_level'] as num).toDouble(),
       memoryUsage: (map['memory_usage'] as num).toDouble(),
-      timestamp: DateTime.now(), // Current time when fetched
+      timestamp: DateTime.now(),
     );
   }
